+++ tests/Doubles/ArticleEditPolicy.php
@@ -5,7 +5,7 @@ namespace Tests\Orisai\Auth\Doubles;
 use Generator;
 use Orisai\Auth\Authentication\Identity;
 use Orisai\Auth\Authorization\AccessEntry;
-use Orisai\Auth\Authorization\AccessEntryType;
+use Orisai\Auth\Authorization\AccessEntryResult;
 use Orisai\Auth\Authorization\Policy;
 use Orisai\Auth\Authorization\PolicyContext;
 
@@ -38,7 +38,7 @@ final class ArticleEditPolicy implements Policy
 			|| $authorizer->isAllowed($identity, ArticleEditOwnedPolicy::getPrivilege(), $requirements);
 
 		yield new AccessEntry(
-			AccessEntryType::fromBool($res),
+			AccessEntryResult::fromBool($res),
 			'',
 		);
 	}