+++ tests/Doubles/PassWithNoIdentityPolicy.php
@@ -5,7 +5,7 @@ namespace Tests\Orisai\Auth\Doubles;
 use Generator;
 use Orisai\Auth\Authentication\Identity;
 use Orisai\Auth\Authorization\AccessEntry;
-use Orisai\Auth\Authorization\AccessEntryType;
+use Orisai\Auth\Authorization\AccessEntryResult;
 use Orisai\Auth\Authorization\NoRequirements;
 use Orisai\Auth\Authorization\OptionalIdentityPolicy;
 use Orisai\Auth\Authorization\PolicyContext;
@@ -29,7 +29,7 @@ final class PassWithNoIdentityPolicy implements OptionalIdentityPolicy
 	public function isAllowed(?Identity $identity, object $requirements, PolicyContext $context): Generator
 	{
 		yield new AccessEntry(
-			AccessEntryType::fromBool($identity === null),
+			AccessEntryResult::fromBool($identity === null),
 			'',
 		);
 	}