+++ tests/Doubles/RequireCurrentUserPolicy.php
@@ -5,7 +5,7 @@ namespace Tests\Orisai\Auth\Doubles;
 use Generator;
 use Orisai\Auth\Authentication\Identity;
 use Orisai\Auth\Authorization\AccessEntry;
-use Orisai\Auth\Authorization\AccessEntryType;
+use Orisai\Auth\Authorization\AccessEntryResult;
 use Orisai\Auth\Authorization\CurrentUserPolicyContext;
 use Orisai\Auth\Authorization\NoRequirements;
 use Orisai\Auth\Authorization\Policy;
@@ -30,7 +30,7 @@ final class RequireCurrentUserPolicy implements Policy
 	public function isAllowed(Identity $identity, object $requirements, PolicyContext $context): Generator
 	{
 		yield new AccessEntry(
-			AccessEntryType::fromBool($context instanceof CurrentUserPolicyContext),
+			AccessEntryResult::fromBool($context instanceof CurrentUserPolicyContext),
 			'',
 		);
 	}