+++ tests/Unit/Authorization/AccessEntryTest.php
@@ -3,7 +3,7 @@
 namespace Tests\Orisai\Auth\Unit\Authorization;
 
 use Orisai\Auth\Authorization\AccessEntry;
-use Orisai\Auth\Authorization\AccessEntryType;
+use Orisai\Auth\Authorization\AccessEntryResult;
 use Orisai\TranslationContracts\TranslatableMessage;
 use PHPUnit\Framework\TestCase;
 
@@ -13,18 +13,18 @@ final class AccessEntryTest extends TestCase
 	public function test(): void
 	{
 		$message = 'Message';
-		$entry = new AccessEntry(AccessEntryType::allowed(), $message);
+		$entry = new AccessEntry(AccessEntryResult::allowed(), $message);
 
-		self::assertSame(AccessEntryType::allowed(), $entry->getType());
+		self::assertSame(AccessEntryResult::allowed(), $entry->getResult());
 		self::assertSame($message, $entry->getMessage());
 	}
 
 	public function testTranslatable(): void
 	{
 		$message = new TranslatableMessage('translatable.message', ['a' => 'b']);
-		$entry = new AccessEntry(AccessEntryType::forbidden(), $message);
+		$entry = new AccessEntry(AccessEntryResult::forbidden(), $message);
 
-		self::assertSame(AccessEntryType::forbidden(), $entry->getType());
+		self::assertSame(AccessEntryResult::forbidden(), $entry->getResult());
 		self::assertSame($message, $entry->getMessage());
 	}